Rename modules and classes to fit the naming scheme.
