+++ index.html
@@ -241,7 +241,7 @@
         <div id="loader">
             <img src="assets/logo.svg" id="spinner">
         </div>
-        <script type="module" src="App.js"></script>
+        <script type="module" src="app.js"></script>
     </main>
 </body>
 </html>
